@@ -1,45 +1,20 @@
 {
   "name": "event-calendar",
-  "version": "3.12.0",
+  "private": true,
+  "version": "4.0.0",
+  "type": "module",
   "scripts": {
-    "build": "rollup -c",
-    "dev": "rollup -c -w",
-    "start": "sirv public",
+    "dev": "vite",
+    "build": "vite build && vite build --mode iife",
+    "preview": "vite preview",
     "set-version": "node ./scripts/set-version.cjs",
     "publish": "node ./scripts/publish.cjs"
   },
   "devDependencies": {
-    "@babel/core": "^7.26.0",
-    "@babel/plugin-transform-runtime": "^7.25.9",
-    "@babel/preset-env": "^7.26.0",
-    "@babel/runtime-corejs3": "^7.26.0",
-    "@event-calendar/build": "~3.12.0",
-    "@event-calendar/core": "~3.12.0",
-    "@event-calendar/day-grid": "~3.12.0",
-    "@event-calendar/interaction": "~3.12.0",
-    "@event-calendar/list": "~3.12.0",
-    "@event-calendar/resource-time-grid": "~3.12.0",
-    "@event-calendar/resource-timeline": "~3.12.0",
-    "@event-calendar/time-grid": "~3.12.0",
-    "@rollup/plugin-babel": "^6.0.4",
-    "@rollup/plugin-commonjs": "^28.0.1",
-    "@rollup/plugin-node-resolve": "^15.3.0",
-    "@rollup/plugin-terser": "^0.4.4",
-    "autoprefixer": "^10.4.20",
-    "cssnano": "^7.0.6",
-    "postcss": "^8.4.49",
-    "rollup": "^4.27.4",
-    "rollup-plugin-livereload": "^2.0.5",
-    "rollup-plugin-scss": "^4.0.0",
-    "rollup-plugin-svelte": "^7.2.2",
-    "sass": "^1.81.0",
-    "svelte": "^4.2.19"
-  },
-  "dependencies": {
-    "sirv-cli": "^2.0.2"
-  },
-  "browserslist": [
-    "defaults and supports fetch"
-  ],
-  "type": "module"
+    "@sveltejs/vite-plugin-svelte": "^5.0.3",
+    "sass": "^1.86.3",
+    "svelte": "^5.27.0",
+    "vite": "^6.2.6",
+    "vite-plugin-banner": "^0.8.0"
+  }
 }
