@@ -1,26 +1,58 @@
 <script>
-    import {getContext} from 'svelte';
-    import {debounce, toISOString, toLocalDate, toViewWithLocalDates, isFunction} from './lib.js';
+    import {getContext, untrack} from 'svelte';
+    import {
+        toISOString, toLocalDate, toViewWithLocalDates, isFunction, task, flushDebounce, hasYScroll
+    } from '#lib';
 
-    let {datesSet, _auxiliary, _activeRange, _queue, _view} = getContext('state');
+    let {
+        datesSet, eventAllUpdated, _auxiliary, _activeRange, _events, _scrollable, _bodyEl, _tasks, _recheckScrollable,
+        _queue, _view
+    } = getContext('state');
 
-    // Set up datesSet callback
-    $: runDatesSet($_activeRange);
+    // datesSet callback
+    $effect(() => {
+        $_activeRange;
+        untrack(() => {
+            if (isFunction($datesSet)) {
+                $datesSet({
+                    start: toLocalDate($_activeRange.start),
+                    end: toLocalDate($_activeRange.end),
+                    startStr: toISOString($_activeRange.start),
+                    endStr: toISOString($_activeRange.end),
+                    view: toViewWithLocalDates($_view)
+                });
+            }
+        });
+    });
 
-    let debounceHandle = {};
-    function runDatesSet(_activeRange) {
-        if (isFunction($datesSet)) {
-            debounce(() => $datesSet({
-                start: toLocalDate(_activeRange.start),
-                end: toLocalDate(_activeRange.end),
-                startStr: toISOString(_activeRange.start),
-                endStr: toISOString(_activeRange.end),
-                view: toViewWithLocalDates($_view)
-            }), debounceHandle, _queue);
-        }
-    }
+    // eventAllUpdated callback
+    $effect(() => {
+        $_events;
+        untrack(() => {
+            if (isFunction($eventAllUpdated)) {
+                task(() => $eventAllUpdated({view: toViewWithLocalDates($_view)}), 'eau', _tasks);
+            }
+        });
+    });
+
+    $effect(() => {
+        $_queue;
+        untrack(() => {
+            flushDebounce($_queue);
+        });
+    });
+
+    $effect(() => {
+        $_recheckScrollable;
+        untrack(() => {
+            if ($_bodyEl) {
+                $_scrollable = hasYScroll($_bodyEl);
+            }
+            $_recheckScrollable = false;
+        });
+    });
 </script>
 
-{#each $_auxiliary as component}
-    <svelte:component this={component} />
+{#each $_auxiliary as Component}
+    <Component/>
 {/each}
