@@ -1,4 +1,7 @@
-import Calendar from './Calendar.svelte';
-
-export {Calendar as default};
-export * from './lib.js';
+export {default as Calendar} from './Calendar.svelte';
+export {default as DayGrid} from './plugins/day-grid';
+export {default as Interaction} from './plugins/interaction';
+export {default as List} from './plugins/list';
+export {default as ResourceTimeGrid} from './plugins/resource-time-grid';
+export {default as ResourceTimeline} from './plugins/resource-timeline';
+export {default as TimeGrid} from './plugins/time-grid';
