@@ -1,18 +1,8 @@
 import {get, writable} from 'svelte/store';
 import {tick} from 'svelte';
-import {createOptions, createParsers} from './options';
-import {
-    activeRange,
-    currentRange,
-    dayGrid,
-    events,
-    now,
-    today,
-    viewDates,
-    viewTitle,
-    view as view2  // hack to avoid a runtime error in SvelteKit dev mode (ReferenceError: view is not defined)
-} from './stores';
-import {keys, intl, intlRange, isFunction, identity} from '../lib.js';
+import {createOptions, createParsers} from './options.js';
+import {activeRange, currentRange, dayGrid, events, now, today, view as view2, viewDates, viewTitle} from './stores.js';
+import {identity, intl, intlRange, isFunction, keys} from '#lib';
 
 export default class {
     constructor(plugins, input) {
@@ -33,7 +23,6 @@ export default class {
 
         // Private stores
         this._queue = writable(new Map());  // debounce queue (beforeUpdate)
-        this._queue2 = writable(new Map());  // debounce queue (afterUpdate)
         this._tasks = new Map();  // timeout IDs for tasks
         this._auxiliary = writable([]);  // auxiliary components
         this._dayGrid = dayGrid(this);
@@ -50,6 +39,7 @@ export default class {
         this._intlTitle = intlRange(this.locale, this.titleFormat);
         this._bodyEl = writable(undefined);
         this._scrollable = writable(false);
+        this._recheckScrollable = writable(false);
         this._viewTitle = viewTitle(this);
         this._viewDates = viewDates(this);
         this._view = view2(this);
@@ -168,7 +158,7 @@ function mergeOpts(...args) {
 
 function filterOpts(opts, state) {
     keys(opts)
-        .filter(key => !validKey(key, state) || key == 'view')
+        .filter(key => !validKey(key, state) || key === 'view')
         .forEach(key => delete opts[key]);
 }
 
