@@ -16,7 +16,7 @@ for (let name of packages) {
     if (name === 'core') {
         const file = dir + '/src/Calendar.svelte';
         const buf = fs.readFileSync(file);
-        const tmp = buf.toString().replace(/\.\/styles\/index\.scss/g, '../index.css');
+        const tmp = buf.toString().replace(/\.\/styles\/index\.scss/g, '../dist/index.css');
 
         fs.writeFileSync(file, tmp);
         publish(dir);
